create MonsterGraph, change Dadguide signatures
DadguideItem and everything that subclasses it now has **kwargs
so that DgMonster can have a graph added to its signature, also
the signatures of query_one, query_many, etc, have all been changed so
that now the DbContext is used instead of the DadguideDatabase for
DgMonster (and everything else). Updated database_context accordingly.