fix handling of evo type checking
this has two parts: (1) I fixed the way I was getting edges from the
graph, since before I was ignoring edge type, which in practice wasn't
an issue probably but in theory could be, and in the future could be,
so it was worth fixing. (2) I also fixed an issue where old data such as
changed evo trees, for example Typhon's, could be interfering with the
evo data returned. I added tstamp as an attribute to EvolutionModel
and am sorting by this when returning an edge in the above method. This
should hopefully "just work" with no further effort needed. Also added
a test case to cover this, and added more documentation about this to
the Dadguide wiki page in the repo.