add material_for to graph, rm from db_context
(1) the numbers for some mats like keepr of gold & dub mythlit disagree
from prod but I'm pretty sure prod is wrong due to having old data mixed
in with current data. I updated the query in the graph to guarantee that
for evos we only have the most recent data. I also tried to check
against padx but padx's data is even more out of date (e.g. keeper of
gold is still listed as a jorm ullr mat) so that wasn't useful at all.
tests for recent monsters without any old data all agree.

(2) this change added a LOT of extra edges incident to some nodes, e.g.
dub mythlit has over 700 edges out of it now. but running id dub mythlit
does not take noticeable time so this is probably not an issue at all.
In the event that it does become an issue in the future we can split the
mat_for data into a separate graph

(3) I didn't create back edges for mat data because they aren't needed
currently and also see (2) and this would double that effect (even if
currently it seems to have no noticeable impact) but we could of course
do that if it seems useful at some point.