Make MonsterModel ready to replace DgMonster fully (#193)
* implement cur, true evo type in graph

* rm true_evo_type from DGMonster, fix revo in MonsterIndex

* move evo mats to graph

this also allowed for removing evo_from and evo_from_id completely
from DgMonster, so that's completely moved to monster_graph now yay

* move prev/next into graph

* remove evo list from DgMonster

all of the queries using this are still very slow because they require
DgMonsters to be constructed, which requires actual db queries and not
just graph queries, so I just moved the slow part outside of DgMonster
and to the DbContext. However removing this method from DgMonster will
make it easier to later replace the entirety of padinfo with
MonsterModels, and when that's done I can stop casting as DgMonsters and
just use MonsterModels from the graph directly, which will be fast so
this was worth it

* add evos to MonsterModel, rm from DgMonster

* move material_of to db_context (not graph though)

* MonsterModels can generate id menus now!

* remove some unused methods from dbcontext

* fix handling of evo type checking

this has two parts: (1) I fixed the way I was getting edges from the
graph, since before I was ignoring edge type, which in practice wasn't
an issue probably but in theory could be, and in the future could be,
so it was worth fixing. (2) I also fixed an issue where old data such as
changed evo trees, for example Typhon's, could be interfering with the
evo data returned. I added tstamp as an attribute to EvolutionModel
and am sorting by this when returning an edge in the above method. This
should hopefully "just work" with no further effort needed. Also added
a test case to cover this, and added more documentation about this to
the Dadguide wiki page in the repo.

* no more database!!!!!!

* remove db_context passing from query_many etc

* put constants decl in try/except to make running tests easier

* add material_for to graph, rm from db_context

(1) the numbers for some mats like keepr of gold & dub mythlit disagree
from prod but I'm pretty sure prod is wrong due to having old data mixed
in with current data. I updated the query in the graph to guarantee that
for evos we only have the most recent data. I also tried to check
against padx but padx's data is even more out of date (e.g. keeper of
gold is still listed as a jorm ullr mat) so that wasn't useful at all.
tests for recent monsters without any old data all agree.

(2) this change added a LOT of extra edges incident to some nodes, e.g.
dub mythlit has over 700 edges out of it now. but running id dub mythlit
does not take noticeable time so this is probably not an issue at all.
In the event that it does become an issue in the future we can split the
mat_for data into a separate graph

(3) I didn't create back edges for mat data because they aren't needed
currently and also see (2) and this would double that effect (even if
currently it seems to have no noticeable impact) but we could of course
do that if it seems useful at some point.

* move build_graph, generate_all_monsters back into init

* fix evo gem query i think

* reformat sql queries

* change evo_model.mats to list(filter()) and add test