Add tree to NamedMonster, tree discovery to id2
Previously there was a note saying not to add any NamedMonster or
NamedMonsterGroup attributes to NamedMonster but I think that's no
longer relevant because MonsterModel doesn't hold a db reference?
So this should be ok?

This specifically fixes a bug that ^id2 dg equip diablos didn't work
because the nickname of diablos was overwritten by ivanus due to
the name change on the specific card, now the entire tree is auto
added to discovery if any single evo in the tree is discovered
via the nickname match.