Fix Batman and Shinrabansho Choco collab id displays (#596)
Fixes an issue where jp & na id numbers of monsters in
these two collabs disagree with each other. We are
displaying the NA number and querying the NA
number, but showing a warning message explaining
what's going on. The JP number is available by
querying the JP number along with the prefix
`jp` attached to the number. If the user querying
is a padglobal admin, an additional warning is displayed
with some help text about the id config gdoc.