Use Emoji class instead of PaneNames (#906)
There was a class "PaneNames" defined for some menus that was now
doing basically the same thing as the new Emoji class, I refactored
slightly to use Emoji instead for the purpose of seeing if emoji were
to be hidden. Even though several menus were touched, the impact is
very minimal and no one should really notice anything different.